test(examples): run BandChat through topology harness
