+++ examples/band-chat/apps/nextjs-betterauth/vitest.config.browser.ts
@@ -10,6 +10,15 @@ import {
   unblockJazzServerNetwork,
 } from "../../../../packages/jazz-tools/tests/browser/testing-server-node.js";
 
+function jazzBrowserTopologyLog(
+  _context: unknown,
+  status: "start" | "complete" | "failed",
+  label: string,
+  elapsedMs: number,
+) {
+  console.info(`[jazz-browser-topology] ${status} ${label} (${elapsedMs}ms)`);
+}
+
 export default defineConfig({
   define: {
     "process.env.NEXT_PUBLIC_JAZZ_APP_ID": JSON.stringify("band-chat-browser-tests"),
@@ -25,6 +34,7 @@ export default defineConfig({
       provider: playwright(),
       instances: [{ browser: "chromium", headless: true }],
       commands: {
+        jazzBrowserTopologyLog,
         jazzServerInfo: async (_context, appId, schema) => jazzServerInfo(appId, schema),
         jazzServerBlockNetwork: async ({ context }, serverUrl) =>
           blockJazzServerNetwork(context, serverUrl),
